requirements.txt
Flask==2.1.2
language-tool-python==0.9.8
pyspellchecker==0.6.2
nltk==3.6.7


